msqdev fixes: sig_update, table updates, Interrupted system call
Fixed a logic in sig_update where it was triggering updates
of the file instead of the ecu.

Fixed a problem with msqdev where the tables were
not being read from their files before checking for changes.

Added the SA_RESTART flag in attempt to resolve the sporadic
"Interrupted system call" error.