adjustments to increase msq-tuner.pl performance
msq-tuner.pl

Added option (-pr) to configure the range of points.
And reduced the default value from 3 to 1 so that it
changes a maximum 9 points instead of 18+.

Along with the recent change in msqdev to retrieve
real time data at a faster rate the time it takes
to complete a cycle is significantly reduced.
A cycle is now completed in a minute or two instead
of ten minutes.