improved serial port error handling
If the MSQSerial destructor is not called the serial
port will not be closed properly.
And subsequent attemps to use the serial port will
result in read errors (returning 0 bytes).

The code has been improved to handle more errors and close
proprly when a SIGTERM or SIGINT is received (such as with Ctrl-C).

How to re-initialize the serial port to overcome the corruption
problem has not be solved yet.