rtdata ',', added DEBUG, serial baud rate fix
Added #define DEBUG to provide debugging output.

Changed rtdata separator from space to , (CSV).

Possibly fixed the serial initialization problem.
When the problem occurred communication was slow
and full of errors and was due to the baud rate
being incorrect.
Since tcsetattr() does not guarantee that all
the options will be set correctly a check has
been added to verify and retry if necessary.