made msq-afr_table function non-linear, added -drpm option
Changed the ve calculation function in msq-afr_table
to use a non-linear function instead of only a linear
function.

Added option to msq-afr_table (-drpm) to limt the rpm above
which no changes should be made in the table.
This is useful when the fit function was built using low
rpm values and when calculated creates absurd high rpm
values.